requirements.txt
flask
langchain==0.2.3
langchain-community==0.2.4
openai==1.34.0
pinecone==4.0.0
pypdf==4.2.0
sentence-transformers==3.0.1
python-dotenv